fix(ui): circular import issue
