+++ invokeai/frontend/web/src/features/nodes/util/graph/generation/buildImagen3Graph.ts
@@ -14,7 +14,7 @@ import {
 } from 'features/nodes/util/graph/graphBuilderUtils';
 import { type GraphBuilderReturn, UnsupportedGenerationModeError } from 'features/nodes/util/graph/types';
 import { t } from 'i18next';
-import { selectMainModelConfig } from 'services/api/endpoints/models';
+import { selectMainModelConfig } from 'features/controlLayers/store/paramsSlice';
 import type { Equals } from 'tsafe';
 import { assert } from 'tsafe';
