+++ invokeai/frontend/web/src/features/nodes/util/graph/generation/buildSD3Graph.ts
@@ -19,7 +19,7 @@ import {
   selectPresetModifiedPrompts,
 } from 'features/nodes/util/graph/graphBuilderUtils';
 import type { GraphBuilderReturn, ImageOutputNodes } from 'features/nodes/util/graph/types';
-import { selectMainModelConfig } from 'services/api/endpoints/models';
+import { selectMainModelConfig } from 'features/controlLayers/store/paramsSlice';
 import type { Invocation } from 'services/api/types';
 import type { Equals } from 'tsafe';
 import { assert } from 'tsafe';
